Use load-grunt-config
Split Gruntfile to file per task configuration
