+++ open-budget/grunt/bower-install.js
@@ -1,0 +1,6 @@
+module.exports = {
+  target: {
+    src: '<%= yeoman.app %>/index.html',
+    exclude: [ "app/bower_components/bootstrap/dist/css/bootstrap.css" ]
+  }
+};
