+++ open-budget/grunt/clean.js
@@ -1,0 +1,13 @@
+module.exports = {
+  dist: {
+    files: [{
+      dot: true,
+      src: [
+        '.tmp',
+        '<%= yeoman.dist %>/*',
+        '!<%= yeoman.dist %>/.git*'
+      ]
+    }]
+  },
+  server: '.tmp'
+};
