+++ open-budget/grunt/concurrent.js
@@ -1,0 +1,10 @@
+module.exports = {
+  dist: [
+    'coffee',
+    'less',
+    'imagemin',
+    'svgmin',
+    'htmlmin',
+  'jst'
+  ]
+};
