+++ open-budget/grunt/mocha.js
@@ -1,0 +1,8 @@
+module.exports = {
+  all: {
+    options: {
+      run: true,
+      urls: ['http://localhost:<%= connect.options.port %>/index.html']
+    }
+  }
+};
