+++ open-budget/grunt/usemin.js
@@ -1,0 +1,7 @@
+module.exports = {
+    html: ['<%= yeoman.dist %>/{,*/}*.html'],
+    css: ['<%= yeoman.dist %>/styles/{,*/}*.css'],
+    options: {
+      dirs: ['<%= yeoman.dist %>']
+    }
+};
