+++ open-budget/grunt/yeoman.js
@@ -1,0 +1,4 @@
+module.exports = {
+  app: require('../bower.json').appPath || 'app',
+  dist: 'dist'
+};
