+++ open-budget/package.json
@@ -19,9 +19,9 @@
     "grunt-contrib-jshint": "~0.6.2",
     "grunt-contrib-jst": "~0.5.1",
     "grunt-contrib-less": "~0.8.1",
+    "grunt-contrib-requirejs": "~0.4.4",
     "grunt-contrib-uglify": "~0.2.2",
     "grunt-contrib-watch": "~0.5.3",
-    "grunt-contrib-requirejs": "~0.4.4",
     "grunt-mocha": "~0.4.0",
     "grunt-modernizr": "~0.4.0",
     "grunt-newer": "~0.6.0",
@@ -30,7 +30,7 @@
     "grunt-usemin": "~0.1.12",
     "jshint-stylish": "~0.1.3",
     "less": "~1.6.0",
-    "load-grunt-tasks": "~0.2.0",
+    "load-grunt-config": "^0.17.2",
     "matchdep": "~0.1.2",
     "requirejs": "^2.1.18",
     "time-grunt": "~0.2.0"
